adjust to new clang-format and clang-tidy config

+++ kernel/interrupts/ioapic.c
@@ -1,9 +1,9 @@
+#include <interrupts/interrupts.h>
+#include <interrupts/ioapic.h>
+#include <interrupts/lapic.h>
+#include <io/io.h>
 #include <memory/pmm.h>
 #include <panic.h>
-#include <interrupts/interrupts.h>
-#include <io/io.h>
-#include <interrupts/lapic.h>
-#include <interrupts/ioapic.h>
 
 #define MAX_IOAPICS 4
 #define IOREGSEL 0x00
@@ -14,9 +14,9 @@
 #define IOAPIC_RENTRY_BASE 0x10
 
 #define IOAPIC_ACTIVE_HIGH 0
-#define IOAPIC_ACTIVE_LOW  1
+#define IOAPIC_ACTIVE_LOW 1
 
-#define IOAPIC_EDGE  0
+#define IOAPIC_EDGE 0
 #define IOAPIC_LEVEL 1
 
 struct ioapic {
@@ -48,19 +48,19 @@ static int nioapics = 0;
 static u8 redirections[256];
 
 static u32 ioapic_read32(void *base_addr, u8 reg) {
-	volatile u32 *regsel = (volatile u32*) (base_addr + IOREGSEL);
-	volatile u32 *regwin = (volatile u32*) (base_addr + IOREGWIN);
-	*regsel = reg;
+	volatile u32 *regsel = (volatile u32 *) (base_addr + IOREGSEL);
+	volatile u32 *regwin = (volatile u32 *) (base_addr + IOREGWIN);
+	*regsel              = reg;
 	return *regwin;
 }
 
 static u64 ioapic_read64(void *base_addr, u8 index) {
-	volatile u32 *regsel = (volatile u32*) (base_addr + IOREGSEL);
-	volatile u32 *regwin = (volatile u32*) (base_addr + IOREGWIN);
+	volatile u32 *regsel = (volatile u32 *) (base_addr + IOREGSEL);
+	volatile u32 *regwin = (volatile u32 *) (base_addr + IOREGWIN);
 	u64 value;
 
 	*regsel = index;
-	value = *regwin;
+	value   = *regwin;
 	*regsel = index + 1;
 	value |= (u64) *regwin << 32;
 
@@ -68,8 +68,8 @@ static u64 ioapic_read64(void *base_addr, u8 index) {
 }
 
 static void ioapic_write64(void *base_addr, u8 index, u64 value) {
-	volatile u32 *regsel = (volatile u32*) (base_addr + IOREGSEL);
-	volatile u32 *regwin = (volatile u32*) (base_addr + IOREGWIN);
+	volatile u32 *regsel = (volatile u32 *) (base_addr + IOREGSEL);
+	volatile u32 *regwin = (volatile u32 *) (base_addr + IOREGWIN);
 
 	*regsel = index;
 	*regwin = value & 0xffffffff;
@@ -83,7 +83,7 @@ static u32 rentry_index(u8 rentry) {
 
 static struct ioapic *ioapic_get_from_irq(u8 irq) {
 	for (int i = 0; i < nioapics; i++) {
-		u8 base = ioapics[i].gsi_base;
+		u8 base    = ioapics[i].gsi_base;
 		u8 entries = ioapics[i].max_rentry;
 		if (irq >= base && irq < base + entries) {
 			return &ioapics[i];
@@ -104,17 +104,17 @@ void ioapic_register(u8 id, u64 addr, u32 gsi_base) {
 		panic("ioapic id %d already registered!", id);
 	}
 
-	ioapic->id = id;
+	ioapic->id   = id;
 	ioapic->addr = pmm_to_virt(addr);
 
-	u32 version = ioapic_read32(ioapic->addr, IOAPIC_VERSION);
+	u32 version        = ioapic_read32(ioapic->addr, IOAPIC_VERSION);
 	ioapic->max_rentry = (version >> 16) & 0xff;
 
 	ioapic->gsi_base = gsi_base;
 
 	kprintf("ioapic id %d, from %d to %d\n", id, gsi_base, gsi_base + ioapic->max_rentry);
 
-	for (u8 i = ioapic->gsi_base; i < ioapic->gsi_base + ioapic->max_rentry; i++) {
+	for (u32 i = ioapic->gsi_base; i < ioapic->gsi_base + ioapic->max_rentry; i++) {
 		ioapic_set_irq_override(i, i, 1);
 		redirections[i] = i;
 	}
@@ -127,10 +127,10 @@ void ioapic_set_irq_override(u8 src_irq, u8 dst_irq, u16 flags) {
 		panic("no ioapic for irq %d", src_irq);
 	}
 
-	union ioapic_rentry rentry = { .raw = ioapic_read64(ioapic->addr, rentry_index(dst_irq)) };
-	rentry.mask = 1;
-	rentry.vector = src_irq + INT_BASE;
-	rentry.dest = apic_ioirq_id();
+	union ioapic_rentry rentry = {.raw = ioapic_read64(ioapic->addr, rentry_index(dst_irq))};
+	rentry.mask                = 1;
+	rentry.vector              = src_irq + INT_BASE;
+	rentry.dest                = apic_ioirq_id();
 
 	u8 polarity = flags & 0x3;
 	if (polarity == 0b00 || polarity == 0b11) {
@@ -167,11 +167,11 @@ void ioapic_set_mask(u8 irq, u8 mask) {
 	}
 
 	if (irq != redirections[irq]) {
-		irq = redirections[irq];
+		irq    = redirections[irq];
 		ioapic = ioapic_get_from_irq(irq);
 	}
 
-	union ioapic_rentry rentry = { .raw = ioapic_read64(ioapic->addr, rentry_index(irq)) };
-	rentry.mask = mask & 1;
+	union ioapic_rentry rentry = {.raw = ioapic_read64(ioapic->addr, rentry_index(irq))};
+	rentry.mask                = mask & 1;
 	ioapic_write64(ioapic->addr, rentry_index(irq), rentry.raw);
 }
